# reports/openapi_day4.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SCM Copilot",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "description": "返回当前登录用户（权限来自 JWT claims，不查库）——受保护端点示例。",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "登录（JWT 双令牌）",
        "description": "bcrypt 校验用户名密码 → 签发 access（15min）+ refresh（24h）双令牌，登录成功/失败均写审计。",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "刷新令牌（rotation）",
        "description": "refresh token 换取新双令牌；旧 refresh 立即吊销（rotation），回放旧 refresh 应 401。",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Err": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "机器可读错误码，如 AUTH_401 / QUOTA_429"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "人类可读错误说明"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "贯穿一次请求的 request_id（排查定位用）"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "Err",
        "description": "统一错误响应体（OpenAPI `responses={401: {\"model\": Err}, ...}` 引用）。"
      },
...
      "LoginIn": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 72,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginIn",
        "description": "登录请求体。\n\n密码长度限制在 bcrypt 72 字节输入上限内（手册坑：超长会 silent truncate，\n登录校验前先拒绝，避免\"短密码能过、长密码被截断成同一 hash\"的时序风险）。"
      },
...
      "RefreshIn": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshIn",
        "description": "refresh 请求体。"
      },
...
      "TokenOut": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenOut",
        "description": "双令牌返回：短效 access + 长效 refresh（泄露窗口 vs 体验的平衡）。"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "tenant_id",
          "permissions"
        ],
        "title": "UserOut",
        "description": "受保护端点回显当前用户（权限来自 JWT claims，不查库）。"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}